requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
gunicorn